workouts: Add the ability to delete a workout.
With the way provider works, we had to remove the workout pop-up
menu button from WorkoutDetailsDialog. As it turns out, you cannot
delete a workout from the context where you are viewing it to
begin with. We replaced the pop-up menu button with a generic edit
button that allows the user to edit the workout from the
WorkoutDetailsDialog.