lib/models: Rename WorkoutCollection -> WorkoutTemplateCollection.
This commit is a part of an overhaul of the app's data model.
Essentially, we want to distinguish between workout "templates"
and workouts "performed". For example, a workout template specifies
the exercises and goal sets and reps. On the other hand, a "workout
performed" object has the actual details on what the user did when
they performed the actual workout. A "workout performed" may have
additional information that the template cannot have. For instance,
a "workout performed" may have fields for duration, PRs and volume.